Switched to string pattern from regexp in feature file
I'm still not sure why the mode and variable setting steps from Espuds
won't work for me but this is better than it was.

Also, strangely, the pattern I changed doesn't seem to match when
I include all of the buffer contents (I'm thinking maybe this has
something to do with the handling of the trailing whitespace). But I was
thinking of moving towards just including the changed line anyway.